feat: change logout to use navigation

+++ todolist/src/utils/logout.ts
@@ -1,5 +1,15 @@
 import { deleteData } from './storage';
+import { Alert } from 'react-native';
 
-export const logout = () => {
-  deleteData('token');
+export const logout = (navigation: any) => {
+  Alert.alert('ออกจากระบบ', 'ท่านต้องการออกจากระบบหรือไม่', [
+    {
+      text: 'ออกจากระบบ',
+      onPress: () => {
+        deleteData('token');
+        navigation.navigate('Login');
+      },
+    },
+    { text: 'ยกเลิก', onPress: () => null },
+  ]);
 };
